requirements.txt
# local package
-e .

# external requirements
pandas
python-dotenv>=0.5.1
requests
os
json
numpy
csv
datetime
dateutil
unicodedata
time
random
argparse
matplotlib